+++ dandelion/module.py
@@ -25,6 +25,7 @@
                              by previous version of Dandelion
                 4, 11, 2018  intercept name change of `Module` class and re-name its parameters accordingly
                 5, 28, 2018  fixed: `convTOP` should be constructed each time the `forward()` function of `ConvTransposed2D` is called.
+                7,  9, 2018  fixed: incorrect `W_shape` when `num_groups` > 1
 
   Note      :
     1) GRU & LSTM and their cell version have built-in activation (tanh), other modules have no built-in activations
@@ -889,7 +890,7 @@ class Conv2D(Module):
         else:                             # pad must be 2-element tuple
             self.pad = pad
 
-        self.W_shape = [out_channels, in_channels, self.kernel_size[0], self.kernel_size[1]]
+        self.W_shape = [out_channels, in_channels//num_groups, self.kernel_size[0], self.kernel_size[1]]
         self.W = self.register_param(W, self.W_shape, name='W_Conv2D')
         if b is not None:
             if untie_bias: